+++ CHANGELOG.md
@@ -36,6 +36,8 @@ Main (unreleased)
 
 - Bump snmp_exporter and embedded modules to 0.27.0. Add support for multi-module handling by comma separation and expose argument to increase SNMP polling concurrency for `prometheus.exporter.snmp`. (@v-zhuravlev)
 
+- Add support for pushv1.PusherService Connect API in `pyroscope.receive_http`. (@simonswine)
+
 - (_Experimental_) Add enable/disable collector configurability to `database_observability.mysql`. This removes the `query_samples_enabled` argument, now configurable via enable/disable collector. (@fridgepoet)
 
 v1.6.1